Politicas

+++ app/Http/Controllers/EmpresasController.php
@@ -6,10 +6,13 @@ use App\Models\Empresa;
 use App\Services\GeocodingService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Validator;
 
 class EmpresasController extends Controller
 {
+    use AuthorizesRequests;
     protected $geocodingService;
 
     public function __construct(GeocodingService $geocodingService)
@@ -47,7 +50,7 @@ class EmpresasController extends Controller
             'telefono' => $request->telefono,
             'direccion' => $request->direccion,
             'codigo_postal' => $request->codigo_postal,
-            'estado_subscripcion' => $request->confirmar_subscripcion ? 'activo' : 'inactivo', // Asegúrate de manejar este campo correctamente
+            'estado_subscripcion' => $request->confirmar_subscripcion ? 'activo' : 'inactivo',
             'coordenadas' => $coordenadas,
             'user_id' => Auth::id(),
         ]);
@@ -59,28 +62,32 @@ class EmpresasController extends Controller
 
     public function show(Empresa $empresa)
     {
+        $this->authorize('view', $empresa);
         return view('empresas.show', compact('empresa'));
     }
 
     public function edit(Empresa $empresa)
     {
+        $this->authorize('update', $empresa);
         return view('empresas.edit', compact('empresa'));
     }
 
     public function update(Request $request, Empresa $empresa)
     {
+        $this->authorize('update', $empresa);
+
         $request->validate([
             'nombre_empresa' => 'required',
             'email' => 'required|email',
             'telefono' => 'required',
             'direccion' => 'required',
             'codigo_postal' => 'required',
-            'estado_subscripcion' => $request->confirmar_subscripcion ? 'activo' : 'inactivo', // Asegúrate de manejar este campo correctamente
+            'estado_subscripcion' => $request->confirmar_subscripcion ? 'activo' : 'inactivo',
         ]);
 
         $coordenadas = $this->geocodingService->getCoordinatesFromAddress($request->direccion . ', ' . $request->codigo_postal);
 
-        $empresa->where('id', $empresa->id)->update([
+        $empresa->update([
             'nombre_empresa' => $request->nombre_empresa,
             'email' => $request->email,
             'telefono' => $request->telefono,
@@ -95,6 +102,7 @@ class EmpresasController extends Controller
 
     public function destroy(Empresa $empresa)
     {
+        $this->authorize('delete', $empresa);
         $empresa->delete();
         return redirect()->route('empresas.index');
     }
